updated for Tumbleweed
Updated the sudoers function to add the USERS_GROUP group with no password to a new /etc/sudoers.d/users include file. Added code to see if the users in the USER_LIST are in the USERS_GROUP and if not add them via usermod -aG. This makes the script compatible with Tumbleweed.