added libguestfs packages
Added libguestfs packages as libguestfs and guestfs-tools have been split out across multiple packages in Leap 15.5. These new packages will fail to install on Leap and SLE older than 15.5 (15-SP5) as they don't exist. This is OK as the same functionality was included in the main guestfs-tools and libguestfs packages in previous versions.